+++ pylabrobot/liquid_handling/backends/serializing_backend_tests.py
@@ -6,7 +6,7 @@ from pylabrobot.resources import (
   STARLetDeck,
   TIP_CAR_480_A00,
   PLT_CAR_L5AC_A00,
-  Cos_96_EZWash,
+  Cor_96_wellplate_360ul_Fb,
   STF_L,
   Coordinate,
   no_tip_tracking,
@@ -29,8 +29,8 @@ class SerializingBackendTests(unittest.IsolatedAsyncioTestCase):
     self.deck.assign_child_resource(self.tip_car, rails=1)
 
     self.plt_car = PLT_CAR_L5AC_A00(name="plate carrier")
-    self.plt_car[0] = self.plate = Cos_96_EZWash(name="plate_01")
-    self.plt_car[1] = self.other_plate = Cos_96_EZWash(name="plate_02")
+    self.plt_car[0] = self.plate = Cor_96_wellplate_360ul_Fb(name="plate_01")
+    self.plt_car[1] = self.other_plate = Cor_96_wellplate_360ul_Fb(name="plate_02")
     self.deck.assign_child_resource(self.plt_car, rails=9)
 
     self.backend.clear()