+++ pylabrobot/liquid_handling/liquid_handler_tests.py
@@ -23,8 +23,7 @@ from pylabrobot.resources import (
   TipRack,
   TIP_CAR_480_A00,
   PLT_CAR_L5AC_A00,
-  Cos_96_DW_1mL,
-  Cos_96_DW_500ul,
+  Cor_96_wellplate_360ul_Fb,
   ResourceNotFoundError,
 )
 from pylabrobot.resources.hamilton import STARLetDeck
@@ -62,8 +61,8 @@ class TestLiquidHandlerLayout(unittest.IsolatedAsyncioTestCase):
     tip_car[3] = HTF_L("tip_rack_04")
 
     plt_car = PLT_CAR_L5AC_A00(name="plate carrier")
-    plt_car[0] = Cos_96_DW_1mL(name="aspiration plate")
-    plt_car[2] = Cos_96_DW_500ul(name="dispense plate")
+    plt_car[0] = Cor_96_wellplate_360ul_Fb(name="aspiration plate")
+    plt_car[2] = Cor_96_wellplate_360ul_Fb(name="dispense plate")
 
     self.deck.assign_child_resource(tip_car, rails=1)
     self.deck.assign_child_resource(plt_car, rails=21)
@@ -93,7 +92,7 @@ class TestLiquidHandlerLayout(unittest.IsolatedAsyncioTestCase):
     tip_car = TIP_CAR_480_A00(name="tip_carrier")
     tip_car[0] = STF_L(name="tip_rack_01")
     plt_car = PLT_CAR_L5AC_A00(name="plate carrier")
-    plt_car[0] = Cos_96_DW_1mL(name="aspiration plate")
+    plt_car[0] = Cor_96_wellplate_360ul_Fb(name="aspiration plate")
     self.deck.assign_child_resource(tip_car, rails=1)
     self.deck.assign_child_resource(plt_car, rails=10)
 
@@ -114,8 +113,8 @@ class TestLiquidHandlerLayout(unittest.IsolatedAsyncioTestCase):
     tip_car[0] = STF_L(name="tip_rack_01")
     tip_car[3] = HTF_L(name="tip_rack_04")
     plt_car = PLT_CAR_L5AC_A00(name="plate carrier")
-    plt_car[0] = Cos_96_DW_1mL(name="aspiration plate")
-    plt_car[2] = Cos_96_DW_500ul(name="dispense plate")
+    plt_car[0] = Cor_96_wellplate_360ul_Fb(name="aspiration plate")
+    plt_car[2] = Cor_96_wellplate_360ul_Fb(name="dispense plate")
     self.deck.assign_child_resource(tip_car, rails=1)
     self.deck.assign_child_resource(plt_car, rails=10)
 
@@ -146,7 +145,7 @@ class TestLiquidHandlerLayout(unittest.IsolatedAsyncioTestCase):
       cast(Plate, self.lh.deck.get_resource("aspiration plate")).get_item("A1")
         .get_absolute_location() +
       cast(Plate, self.lh.deck.get_resource("aspiration plate")).get_item("A1").center(),
-        Coordinate(320.500, 146.000, 187.150))
+        Coordinate(x=320.8, y=145.7, z=189.18) )
 
   def test_illegal_subresource_assignment_before(self):
     # Test assigning subresource with the same name as another resource in another carrier. This
@@ -154,7 +153,7 @@ class TestLiquidHandlerLayout(unittest.IsolatedAsyncioTestCase):
     tip_car = TIP_CAR_480_A00(name="tip_carrier")
     tip_car[0] = STF_L(name="sub")
     plt_car = PLT_CAR_L5AC_A00(name="plate carrier")
-    plt_car[0] = Cos_96_DW_1mL(name="sub")
+    plt_car[0] = Cor_96_wellplate_360ul_Fb(name="sub")
     self.deck.assign_child_resource(tip_car, rails=1)
     with self.assertRaises(ValueError):
       self.deck.assign_child_resource(plt_car, rails=10)
@@ -165,15 +164,15 @@ class TestLiquidHandlerLayout(unittest.IsolatedAsyncioTestCase):
     tip_car = TIP_CAR_480_A00(name="tip_carrier")
     tip_car[0] = STF_L(name="sub")
     plt_car = PLT_CAR_L5AC_A00(name="plate carrier")
-    plt_car[0] = Cos_96_DW_1mL(name="ok")
+    plt_car[0] = Cor_96_wellplate_360ul_Fb(name="ok")
     self.deck.assign_child_resource(tip_car, rails=1)
     self.deck.assign_child_resource(plt_car, rails=10)
     with self.assertRaises(ValueError):
-      plt_car[1] = Cos_96_DW_500ul(name="sub")
+      plt_car[1] = Cor_96_wellplate_360ul_Fb(name="sub")
 
   async def test_move_plate_to_site(self):
     plt_car = PLT_CAR_L5AC_A00(name="plate carrier")
-    plt_car[0] = plate = Cos_96_DW_1mL(name="plate")
+    plt_car[0] = plate = Cor_96_wellplate_360ul_Fb(name="plate")
     self.deck.assign_child_resource(plt_car, rails=21)
 
     await self.lh.move_plate(plate, plt_car[2])
@@ -181,11 +180,11 @@ class TestLiquidHandlerLayout(unittest.IsolatedAsyncioTestCase):
     self.assertIsNone(plt_car[0].resource)
     self.assertEqual(plt_car[2].resource, self.lh.deck.get_resource("plate"))
     self.assertEqual(plate.get_item("A1").get_absolute_location() + plate.get_item("A1").center(),
-                     Coordinate(568.000, 338.000, 187.150))
+                     Coordinate(x=568.3, y=337.7, z=189.18))
 
   async def test_move_plate_free(self):
     plt_car = PLT_CAR_L5AC_A00(name="plate carrier")
-    plt_car[0] = plate = Cos_96_DW_1mL(name="plate")
+    plt_car[0] = plate = Cor_96_wellplate_360ul_Fb(name="plate")
     self.deck.assign_child_resource(plt_car, rails=1)
 
     await self.lh.move_plate(plate, Coordinate(1000, 1000, 1000))
@@ -232,7 +231,7 @@ class TestLiquidHandlerCommands(unittest.IsolatedAsyncioTestCase):
     self.lh = LiquidHandler(backend=self.backend, deck=self.deck)
 
     self.tip_rack = STF_L(name="tip_rack")
-    self.plate = Cos_96_DW_1mL(name="plate")
+    self.plate = Cor_96_wellplate_360ul_Fb(name="plate")
     self.deck.assign_child_resource(self.tip_rack, location=Coordinate(0, 0, 0))
     self.deck.assign_child_resource(self.plate, location=Coordinate(100, 100, 0))
     await self.lh.setup()
@@ -622,7 +621,7 @@ class TestLiquidHandlerVolumeTracking(unittest.IsolatedAsyncioTestCase):
     self.deck = STARLetDeck()
     self.lh = LiquidHandler(backend=self.backend, deck=self.deck)
     self.tip_rack = STF_L(name="tip_rack")
-    self.plate = Cos_96_DW_1mL(name="plate")
+    self.plate = Cor_96_wellplate_360ul_Fb(name="plate")
     self.deck.assign_child_resource(self.tip_rack, location=Coordinate(0, 0, 0))
     self.deck.assign_child_resource(self.plate, location=Coordinate(100, 100, 0))
     await self.lh.setup()
@@ -658,7 +657,7 @@ class TestLiquidHandlerCrossContaminationTracking(unittest.IsolatedAsyncioTestCa
     self.deck = STARLetDeck()
     self.lh = LiquidHandler(backend=self.backend, deck=self.deck)
     self.tip_rack = STF_L(name="tip_rack")
-    self.plate = Cos_96_DW_1mL(name="plate")
+    self.plate = Cor_96_wellplate_360ul_Fb(name="plate")
     self.deck.assign_child_resource(self.tip_rack, location=Coordinate(0, 0, 0))
     self.deck.assign_child_resource(self.plate, location=Coordinate(100, 100, 0))
     await self.lh.setup()