+++ pylabrobot/resources/hamilton/hamilton_decks.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from abc import ABCMeta, abstractmethod
-import inspect
 import logging
 from typing import Optional, cast
 
@@ -11,7 +10,6 @@ from pylabrobot.resources.deck import Deck
 from pylabrobot.resources.resource import Resource
 from pylabrobot.resources.trash import Trash
 from pylabrobot.resources.ml_star.mfx_modules import MFXModule
-import pylabrobot.utils.file_parsing as file_parser
 
 
 logger = logging.getLogger("pylabrobot")
@@ -150,7 +148,9 @@ class HamiltonDeck(Deck, metaclass=ABCMeta):
 
   @classmethod
   def load_from_lay_file(cls, fn: str) -> HamiltonDeck:
-    """ Parse a .lay file (legacy layout definition) and build the layout on this deck.
+    """ [DEPRECATED]
+
+    Parse a .lay file (legacy layout definition) and build the layout on this deck.
 
     Args:
       fn: Filename of .lay file.
@@ -163,81 +163,83 @@ class HamiltonDeck(Deck, metaclass=ABCMeta):
       >>> deck = HamiltonSTARDeck.load_from_lay_file("deck.lay")
     """
 
-    # pylint: disable=import-outside-toplevel, cyclic-import
-    import pylabrobot.resources as resources_module
+    raise NotImplementedError("This method is deprecated and will be removed in a future release.")
 
-    c = None
-    with open(fn, "r", encoding="ISO-8859-1") as f:
-      c = f.read()
+    # # pylint: disable=import-outside-toplevel, cyclic-import
+    # import pylabrobot.resources as resources_module
 
-    deck_type = file_parser.find_string("Deck", c)
+    # c = None
+    # with open(fn, "r", encoding="ISO-8859-1") as f:
+    #   c = f.read()
 
-    num_rails = {"ML_Starlet.dck": STARLET_NUM_RAILS, "ML_STAR2.deck": STAR_NUM_RAILS}[deck_type]
-    size_x = {"ML_Starlet.dck": STARLET_SIZE_X, "ML_STAR2.deck": STAR_SIZE_X}[deck_type]
-    size_y = {"ML_Starlet.dck": STARLET_SIZE_Y, "ML_STAR2.deck": STAR_SIZE_Y}[deck_type]
-    size_z = {"ML_Starlet.dck": STARLET_SIZE_Z, "ML_STAR2.deck": STAR_SIZE_Z}[deck_type]
+    # deck_type = file_parser.find_string("Deck", c)
 
-    deck = cls(num_rails=num_rails,
-      size_x=size_x, size_y=size_y, size_z=size_z,
-      origin=Coordinate.zero())
+    # num_rails = {"ML_Starlet.dck": STARLET_NUM_RAILS, "ML_STAR2.deck": STAR_NUM_RAILS}[deck_type]
+    # size_x = {"ML_Starlet.dck": STARLET_SIZE_X, "ML_STAR2.deck": STAR_SIZE_X}[deck_type]
+    # size_y = {"ML_Starlet.dck": STARLET_SIZE_Y, "ML_STAR2.deck": STAR_SIZE_Y}[deck_type]
+    # size_z = {"ML_Starlet.dck": STARLET_SIZE_Z, "ML_STAR2.deck": STAR_SIZE_Z}[deck_type]
 
-    # Get class names of all defined resources.
-    resource_classes = [c[0] for c in inspect.getmembers(resources_module)]
+    # deck = cls(num_rails=num_rails,
+    #   size_x=size_x, size_y=size_y, size_z=size_z,
+    #   origin=Coordinate.zero())
 
-    # Get number of items on deck.
-    num_items = file_parser.find_int("Labware.Cnt", c)
+    # # Get class names of all defined resources.
+    # resource_classes = [c[0] for c in inspect.getmembers(resources_module)]
 
-    # Collect all items on deck.
+    # # Get number of items on deck.
+    # num_items = file_parser.find_int("Labware.Cnt", c)
 
-    containers = {}
-    children = {}
+    # # Collect all items on deck.
 
-    for i in range(1, num_items+1):
-      name = file_parser.find_string(f"Labware.{i}.Id", c)
+    # containers = {}
+    # children = {}
 
-      # get class name (generated from file name)
-      file_name = file_parser.find_string(f"Labware.{i}.File", c).split("\\")[-1]
-      class_name = None
-      if ".rck" in file_name:
-        class_name = file_name.split(".rck")[0]
-      elif ".tml" in file_name:
-        class_name = file_name.split(".tml")[0]
+    # for i in range(1, num_items+1):
+    #   name = file_parser.find_string(f"Labware.{i}.Id", c)
 
-      if class_name in resource_classes:
-        klass = getattr(resources_module, class_name)
-        resource = klass(name=name)
-      else:
-        logger.warning(
-          "Resource with classname %s not found. Please file an issue at "
-          "https://github.com/pylabrobot/pylabrobot/issues/new?assignees=&labels="
-          "&title=Deserialization%%3A%%20Class%%20%s%%20not%%20found", class_name, class_name)
-        continue
+    #   # get class name (generated from file name)
+    #   file_name = file_parser.find_string(f"Labware.{i}.File", c).split("\\")[-1]
+    #   class_name = None
+    #   if ".rck" in file_name:
+    #     class_name = file_name.split(".rck")[0]
+    #   elif ".tml" in file_name:
+    #     class_name = file_name.split(".tml")[0]
 
-      # get location props
-      # 'default' template means resource are placed directly on the deck, otherwise it
-      # contains the name of the containing resource.
-      if file_parser.find_string(f"Labware.{i}.Template", c) == "default":
-        x = file_parser.find_float(f"Labware.{i}.TForm.3.X", c)
-        y = file_parser.find_float(f"Labware.{i}.TForm.3.Y", c)
-        z = file_parser.find_float(f"Labware.{i}.ZTrans", c)
-        resource.location = Coordinate(x=x, y=y, z=z)
-        containers[name] = resource
-      else:
-        children[name] = {
-          "container": file_parser.find_string(f"Labware.{i}.Template", c),
-          "site": file_parser.find_int(f"Labware.{i}.SiteId", c),
-          "resource": resource}
+    #   if class_name in resource_classes:
+    #     klass = getattr(resources_module, class_name)
+    #     resource = klass(name=name)
+    #   else:
+    #     logger.warning(
+    #       "Resource with classname %s not found. Please file an issue at "
+    #       "https://github.com/pylabrobot/pylabrobot/issues/new?assignees=&labels="
+    #       "&title=Deserialization%%3A%%20Class%%20%s%%20not%%20found", class_name, class_name)
+    #     continue
 
-    # Assign all containers to the deck.
-    for cont in containers.values():
-      deck.assign_child_resource(cont, location=cont.location)
+    #   # get location props
+    #   # 'default' template means resource are placed directly on the deck, otherwise it
+    #   # contains the name of the containing resource.
+    #   if file_parser.find_string(f"Labware.{i}.Template", c) == "default":
+    #     x = file_parser.find_float(f"Labware.{i}.TForm.3.X", c)
+    #     y = file_parser.find_float(f"Labware.{i}.TForm.3.Y", c)
+    #     z = file_parser.find_float(f"Labware.{i}.ZTrans", c)
+    #     resource.location = Coordinate(x=x, y=y, z=z)
+    #     containers[name] = resource
+    #   else:
+    #     children[name] = {
+    #       "container": file_parser.find_string(f"Labware.{i}.Template", c),
+    #       "site": file_parser.find_int(f"Labware.{i}.SiteId", c),
+    #       "resource": resource}
 
-    # Assign child resources to their parents.
-    for child in children.values():
-      cont = containers[child["container"]]
-      cont[5 - child["site"]] = child["resource"]
+    # # Assign all containers to the deck.
+    # for cont in containers.values():
+    #   deck.assign_child_resource(cont, location=cont.location)
 
-    return deck
+    # # Assign child resources to their parents.
+    # for child in children.values():
+    #   cont = containers[child["container"]]
+    #   cont[5 - child["site"]] = child["resource"]
+
+    # return deck
 
   def summary(self) -> str:
     """ Return a summary of the deck.